requirements.txt
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.11.4
finnhub-python==2.4.24
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multitasking==0.0.11
numpy==2.3.1
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.2
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
